feat: :sparkles: crud documentTypes

+++ vuelosglobales/src/main/java/com/vuelosglobales/airports/adapters/out/AirportsMySQLRepository.java
@@ -5,8 +5,8 @@ import java.sql.DriverManager;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.ArrayList;
 import java.util.List;
+import java.util.ArrayList;
 import java.util.Optional;
 
 import com.vuelosglobales.airports.domain.models.Airports;
@@ -79,14 +79,37 @@ public class AirportsMySQLRepository implements AirportsRepository{
     @Override
     public void delete(String id) {
         try (Connection connection = DriverManager.getConnection(url, user, password)) {
-            String query = "DELETE FROM airports WHERE id = ?";
-            try (PreparedStatement statement = connection.prepareStatement(query)) {
-                statement.setString(1, id);
-                statement.executeUpdate();
+            String query = "SELECT COUNT(*) FROM flight_connections WHERE id_origin = ? OR id_destination = ?";
+            try (PreparedStatement stmt = connection.prepareStatement(query)) {
+                stmt.setString(1, id);
+                stmt.setString(2, id);
+                try (ResultSet resultSet = stmt.executeQuery()) {
+                    if (resultSet.next()) {
+                        if (resultSet.getInt(1) > 0) {
+                            System.out.println("[¡]ERROR. En este momento no se puede borrar el aeropuerto porque hay vuelos relacionados a este");
+                        } else {
+                            String query1 = "UPDATE employees SET id_airport = NULL WHERE id_airport = ?";
+                            try (PreparedStatement statement = connection.prepareStatement(query1)) {
+                                statement.setString(1, id);
+                                statement.executeUpdate();
+                            }
+                            String query2 = "UPDATE gates SET id_airport = NULL WHERE id_airport = ?";
+                            try (PreparedStatement statement = connection.prepareStatement(query2)) {
+                                statement.setString(1, id);
+                                statement.executeUpdate();
+                            }
+                            String query3 = "DELETE FROM airports WHERE id = ?";
+                            try (PreparedStatement statement = connection.prepareStatement(query3)) {
+                                statement.setString(1, id);
+                                statement.executeUpdate();
+                            }
+                        }
+                    }
+                }
             }
         } catch (SQLException e) {
             e.printStackTrace();
-        }
+        }        
     }
 
     @Override
